fix: track of profile and super properties of contacts
