+++ packages/pn-personafisica-webapp/src/utility/MixpanelUtils/Strategies/SendAddAddressStrategy.ts
@@ -1,0 +1,71 @@
+import { EventPropertyType, EventStrategy, TrackedEvent } from '@pagopa-pn/pn-commons';
+
+import { CourtesyChannelType, DigitalAddress, LegalChannelType } from '../../../models/contacts';
+import { SaveDigitalAddressParams } from '../../../redux/contact/types';
+
+type SendAddAddressReturn =
+  | {
+      SEND_HAS_EMAIL: 'yes';
+    }
+  | {
+      SEND_HAS_SMS: 'yes';
+    }
+  | {
+      SEND_HAS_PEC: 'yes';
+    };
+
+type SendAddAddressData = {
+  payload: DigitalAddress | void;
+  params: SaveDigitalAddressParams;
+};
+
+export class SendAddAddressStrategy implements EventStrategy {
+  performComputations({ payload, params }: SendAddAddressData): TrackedEvent<SendAddAddressReturn> {
+    const { channelType } = params;
+    // Check payload to distinguish between the action called before PIN validation and after
+    // We have to track only the action after the PIN validation
+    if (!payload || params.senderId !== 'default') {
+      return {};
+    }
+
+    if (channelType === LegalChannelType.PEC && !(payload && payload.pecValid)) {
+      return {};
+    }
+
+    if (channelType === CourtesyChannelType.EMAIL) {
+      return {
+        [EventPropertyType.PROFILE]: {
+          SEND_HAS_EMAIL: 'yes',
+        },
+        [EventPropertyType.SUPER_PROPERTY]: {
+          SEND_HAS_EMAIL: 'yes',
+        },
+      };
+    }
+
+    if (channelType === CourtesyChannelType.SMS) {
+      return {
+        [EventPropertyType.PROFILE]: {
+          SEND_HAS_SMS: 'yes',
+        },
+        [EventPropertyType.SUPER_PROPERTY]: {
+          SEND_HAS_SMS: 'yes',
+        },
+      };
+    }
+
+    if (channelType === LegalChannelType.PEC) {
+      return {
+        [EventPropertyType.PROFILE]: {
+          SEND_HAS_PEC: 'yes',
+        },
+        [EventPropertyType.SUPER_PROPERTY]: {
+          SEND_HAS_PEC: 'yes',
+        },
+      };
+    }
+
+    // AppIO case
+    return {};
+  }
+}
